+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "Background Phi Colors" extension will be documented 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 1.0.0 - 2019-04-??
+
+### Changed
+
+- Changed `background-phi-colors.showIndentErrorInOverviewRulerLane` setting name to `background-phi-colors.showIndentErrorInOverviewRuler`.
+- Changed `background-phi-colors.showActiveTokenInOverviewRulerLane` setting name to `background-phi-colors.showActiveTokenInOverviewRuler`.
+- Changed `background-phi-colors.showTrailingSpacesErrorInOverviewRulerLane` setting name to `background-phi-colors.showTrailingSpacesErrorInOverviewRuler`.
+
 ## 0.8.0 - 2019-04-11
 
 ### Changed